requirements.txt
Flask==2.3.2
pytesseract==0.3.10
Pillow==10.0.0
pdf2image==1.16.3
gunicorn==21.2.0
